Take a pass at factoring out Auto.
